symlogos.py
from sympy import symbols
from sympy.logic import *

# Define basic classes for propositional variables and modal operators.

class Proposition:
    def __init__(self, name):
        self.name = name

    def __str__(self):
        return self.name

    def __repr__(self):
        return f"Proposition('{self.name}')"

    def not_(self):
        return Not(self)

    def box(self):
        return Necessity(self)

    def diamond(self):
        return Possibility(self)


class And:
    def __init__(self, left, right):
        self.left = left
        self.right = right

    def __str__(self):
        left_str = str(self.left)
        right_str = str(self.right)
        return f"({left_str} ∧ {right_str})"

    def __repr__(self):
        return f"And({repr(self.left)}, {repr(self.right)})"

class Not:
    def __init__(self, formula):
        self.formula = formula

    def __str__(self):
        return f"¬{self.formula}"

    def __repr__(self):
        return f"Not({repr(self.formula)})"


class Necessity:
    def __init__(self, formula):
        self.formula = formula

    def __str__(self):
        return f"□{self.formula}"

    def __repr__(self):
        return f"Necessity({repr(self.formula)})"


class Possibility:
    def __init__(self, formula):
        self.formula = formula

    def __str__(self):
        return f"◇{self.formula}"

    def __repr__(self):
        return f"Possibility({repr(self.formula)})"

# Add functions for defining axioms, proving theorems, and checking consistency and validity.

def define_axiom():
    pass

def prove_theorem():
    pass

def check_consistency():
    pass

def check_validity():
    pass

# Example usage:

if __name__ == "__main__":
    p = Proposition("p")
    not_p = p.neg()
    box_p = p.box()
    diamond_p = p.diamond()

    print(not_p)
    print(box_p)
    print(diamond_p)
